Feature flag assistant endpoint + opt in UI
