+++ apps/studio/components/interfaces/Organization/GeneralSettings/AIOptInLevelSelector.tsx
@@ -15,6 +15,21 @@ interface AIOptInLevelSelectorProps {
   layout?: 'horizontal' | 'vertical' | 'flex-row-reverse'
 }
 
+const AI_OPT_IN_LEVELS_OLD = [
+  {
+    value: 'disabled',
+    title: 'Disabled',
+    description:
+      'You do not consent to sharing any database information with Supabase AI and understand that responses will be generic and not tailored to your database',
+  },
+  {
+    value: 'schema',
+    title: 'Send anonymous metadata',
+    description:
+      'You consent to sending anonymous data to Supabase AI, which can improve the answers it shows you.',
+  },
+]
+
 const AI_OPT_IN_LEVELS = [
   {
     value: 'disabled',
@@ -49,30 +64,37 @@ export const AIOptInLevelSelector = ({
   layout = 'vertical',
 }: AIOptInLevelSelectorProps) => {
   const newOrgAiOptIn = useFlag('newOrgAiOptIn')
+  const useBedrockAssistant = useFlag('useBedrockAssistant')
+
+  const optInLevels = useBedrockAssistant ? AI_OPT_IN_LEVELS : AI_OPT_IN_LEVELS_OLD
 
   return (
     <FormItemLayout
       label={label}
+      layout={layout}
       description={
         <div className="flex flex-col gap-y-4 my-4 max-w-xl">
-          {!newOrgAiOptIn && (
-            <Admonition
-              type="note"
-              title="Assistant Opt-in is temporarily disabled"
-              description="We will re-enable opting in to the assistant shortly!"
-            />
+          {useBedrockAssistant && (
+            <>
+              {!newOrgAiOptIn && (
+                <Admonition
+                  type="note"
+                  title="Assistant Opt-in is temporarily disabled"
+                  description="We will re-enable opting in to the assistant shortly!"
+                />
+              )}
+              <p>
+                Supabase AI can provide more relevant answers if you choose to share different
+                levels of data. This feature is powered by Amazon Bedrock which does not store or
+                log your prompts and completions, nor does it use them to train AWS models or
+                distribute them to third parties. This is an organization-wide setting, so please
+                select the level of data you are comfortable sharing.
+              </p>
+            </>
           )}
-          <p>
-            Supabase AI can provide more relevant answers if you choose to share different levels of
-            data. This feature is powered by Amazon Bedrock which does not store or log your prompts
-            and completions, nor does it use them to train AWS models or distribute them to third
-            parties. This is an organization-wide setting, so please select the level of data you
-            are comfortable sharing.
-          </p>
           <OptInToOpenAIToggle />
         </div>
       }
-      layout={layout}
     >
       <div className="max-w-xl">
         <FormField_Shadcn_
@@ -85,7 +107,7 @@ export const AIOptInLevelSelector = ({
               disabled={disabled}
               className="space-y-2 mb-6"
             >
-              {AI_OPT_IN_LEVELS.map((item) => (
+              {optInLevels.map((item) => (
                 <div key={item.value} className="flex items-start space-x-3">
                   <RadioGroupItem_Shadcn_
                     value={item.value}
