+++ apps/studio/hooks/ui/useFlag.ts
@@ -13,6 +13,8 @@ export function useFlag<T = boolean>(name: string) {
 
   const store = flagStore.configcat
 
+  if (name === 'useBedrockAssistant') return false
+
   if (!isObjectEmpty(store) && store[name] === undefined) {
     console.error(`Flag key "${name}" does not exist in ConfigCat flag store`)
     return false
