+++ apps/studio/middleware.ts
@@ -8,6 +8,7 @@ export const config = {
 // [Joshen] Return 404 for all next.js API endpoints EXCEPT the ones we use in hosted:
 const HOSTED_SUPPORTED_API_URLS = [
   '/ai/sql/generate-v3',
+  '/ai/sql/generate-v4',
   '/ai/edge-function/complete',
   '/ai/onboarding/design',
   '/ai/sql/complete',
